use DocumentMove component for bulk and single move
