+++ shared/i18n/locales/en_US/translation.json
@@ -195,16 +195,6 @@
   "Are you sure you want to delete <em>{{ count }} documents</em>? This action will move them to the trash._plural": "Are you sure you want to delete <em>{{ count }} documents</em>? This action will move them to the trash.",
   "Deleting": "Deleting",
   "I'm sure – Delete": "I'm sure – Delete",
-  "Select a location to move": "Select a location to move",
-  "{{ count }} documents moved": "{{ count }} documents moved",
-  "{{ count }} documents moved_plural": "{{ count }} documents moved",
-  "{{ successCount }} moved, {{ errorCount }} failed": "{{ successCount }} moved, {{ errorCount }} failed",
-  "Couldn't move the documents, try again?": "Couldn't move the documents, try again?",
-  "Move {{ count }} documents to <em>{{ location }}</em>": "Move {{ count }} documents to <em>{{ location }}</em>",
-  "Untitled": "Untitled",
-  "Select a location to move {{ count }} documents": "Select a location to move {{ count }} documents",
-  "Select a location to move {{ count }} documents_plural": "Select a location to move {{ count }} documents",
-  "Moving": "Moving",
   "Archive {{ count }} documents": "Archive {{ count }} documents",
   "Archive {{ count }} documents_plural": "Archive {{ count }} documents",
   "Delete {{ count }} documents": "Delete {{ count }} documents",
@@ -242,11 +232,13 @@
   "no access": "no access",
   "You do not have permission to move {{ documentName }} to the {{ collectionName }} collection": "You do not have permission to move {{ documentName }} to the {{ collectionName }} collection",
   "Move document": "Move document",
+  "Moving": "Moving",
   "Moving the document <em>{{ title }}</em> to the {{ newCollectionName }} collection will change permission for all workspace members from <em>{{ prevPermission }}</em> to <em>{{ newPermission }}</em>.": "Moving the document <em>{{ title }}</em> to the {{ newCollectionName }} collection will change permission for all workspace members from <em>{{ prevPermission }}</em> to <em>{{ newPermission }}</em>.",
   "Collections could not be loaded, please reload the app": "Collections could not be loaded, please reload the app",
   "Start view": "Start view",
   "Install now": "Install now",
   "Deleted Collection": "Deleted Collection",
+  "Untitled": "Untitled",
   "Unpin": "Unpin",
   "Select a location to copy": "Select a location to copy",
   "Document copied": "Document copied",
@@ -810,9 +802,16 @@
   "Are you sure about that? Deleting the <em>{{ documentTitle }}</em> document will delete all of its history and <em>{{ any }} nested document</em>.": "Are you sure about that? Deleting the <em>{{ documentTitle }}</em> document will delete all of its history and <em>{{ any }} nested document</em>.",
   "Are you sure about that? Deleting the <em>{{ documentTitle }}</em> document will delete all of its history and <em>{{ any }} nested document</em>._plural": "Are you sure about that? Deleting the <em>{{ documentTitle }}</em> document will delete all of its history and <em>{{ any }} nested document</em>.",
   "If you’d like the option of referencing or restoring the {{noun}} in the future, consider archiving it instead.": "If you’d like the option of referencing or restoring the {{noun}} in the future, consider archiving it instead.",
+  "Select a location to move": "Select a location to move",
   "Document moved": "Document moved",
+  "{{ count }} documents moved": "{{ count }} documents moved",
+  "{{ count }} documents moved_plural": "{{ count }} documents moved",
+  "{{ successCount }} moved, {{ errorCount }} failed": "{{ successCount }} moved, {{ errorCount }} failed",
   "Couldn’t move the document, try again?": "Couldn’t move the document, try again?",
+  "Move {{ count }} documents to <em>{{ location }}</em>": "Move {{ count }} documents to <em>{{ location }}</em>",
   "Move to <em>{{ location }}</em>": "Move to <em>{{ location }}</em>",
+  "Select a location to move {{ count }} documents": "Select a location to move {{ count }} documents",
+  "Select a location to move {{ count }} documents_plural": "Select a location to move {{ count }} documents",
   "Couldn’t create the document, try again?": "Couldn’t create the document, try again?",
   "Document permanently deleted": "Document permanently deleted",
   "Are you sure you want to permanently delete the <em>{{ documentTitle }}</em> document? This action is immediate and cannot be undone.": "Are you sure you want to permanently delete the <em>{{ documentTitle }}</em> document? This action is immediate and cannot be undone.",
